builtin: make v_realloc use int parameter instead of u32

+++ vlib/builtin/array.v
@@ -96,7 +96,7 @@ fn (mut a array) ensure_cap(required int) {
 	if a.cap == 0 {
 		a.data = vcalloc(cap * a.element_size)
 	} else {
-		a.data = v_realloc(a.data, u32(cap * a.element_size))
+		a.data = v_realloc(a.data, cap * a.element_size)
 	}
 	a.cap = cap
 }
